Review fixes

+++ action/deregister.go
@@ -2,6 +2,7 @@ package action
 
 import (
 	"errors"
+	"fmt"
 	"io/ioutil"
 	"strconv"
 
@@ -22,14 +23,18 @@ const (
 // DeregisterCLI removes a backend from VaaS using CLI data
 func DeregisterCLI(c *cli.Context) error {
 	config := getCommonParameters(c.Parent().Parent())
-	debug(config)
+	log.Debugf("Deregister CLI config: %+v\n", config)
+
+	if config.Director == "" {
+		return errors.New("no VaaS director specified")
+	}
 
 	apiClient := vaas.NewClient(config.VaaSURL, config.VaaSUser, config.VaaSKey)
 	backendID := c.Int(FlagBackendID)
 
 	if backendID != 0 {
 		if err := apiClient.DeleteBackend(backendID); err != nil {
-			return err
+			return fmt.Errorf("could not deregister: %s", err)
 		}
 
 		log.WithField(FlagBackendID, backendID).
@@ -40,14 +45,20 @@ func DeregisterCLI(c *cli.Context) error {
 
 // DeregisterK8s configures a VaaS client from K8s data and removes a backend
 func DeregisterK8s(podInfo *k8s.PodInfo, config CommonConfig) error {
+	director, err := podInfo.GetDirector()
+	if err != nil {
+		return fmt.Errorf("could not find VaaS director in Pod info: %s", err)
+	}
+
 	apiClient := vaas.NewClient(config.VaaSURL, config.VaaSUser, config.VaaSKey)
 
 	backendID, err := loadBackendID()
 	if err != nil {
-		return err
+		return fmt.Errorf("could not determine backend ID: %s", err)
 	}
+	log.Infof("Deregistering backend %d from director %s", backendID, director)
 	if err := apiClient.DeleteBackend(backendID); err != nil {
-		return err
+		return fmt.Errorf("could not deregister: %s", err)
 	}
 
 	return errors.New("backend ID not provided")
